Add fallback when Nix path does not exist
We do not always know at build-time where we will get an executable.
For instance, on Nix-less systems, the Nix store will be unavailable
and we need to get it some other way. We could hardcode
/usr/local/bin/ to PATH so that GHC will find it, but this introduces
unnecessary impurities and requires changing hashes between builds. As
a workaround, just use classic “which” functionality.